chore(models-http-api): use rate_limit mod in embedding
