+++ crates/http-api-bindings/Cargo.toml
@@ -19,6 +19,7 @@ tabby-inference = { path = "../tabby-inference" }
 ollama-api-bindings = { path = "../ollama-api-bindings" }
 async-openai.workspace = true
 ratelimit = "0.10"
+tokio.workspace = true
 
 [dev-dependencies]
 tokio = { workspace = true, features = ["rt", "macros"] }
